Added New and Edit Review feature. Added security features to prevent users accessing and invoking actions on other users data

+++ app/controllers/users_controller.rb
@@ -1,11 +1,16 @@
 class UsersController < ApplicationController
   def new
-    @user = User.new
+    # To prevent users already logged in to access sign up page directly
+    unless @current_user.present?
+      @user = User.new
+    else
+      redirect_to root_path
+    end
   end
 
   def create
     @user = User.new user_params
-    
+
     if @user.save
       session[:user_id] = @user.id
       redirect_to root_path
@@ -16,29 +21,35 @@ class UsersController < ApplicationController
 
   def show
     @user = User.find(params[:id])
-    
   end
 
   def edit
-    @user = User.find(params[:id])
-    
+    # To prevent users editing other user profiles
+    if @current_user.id == params[:id].to_i
+      @user = User.find(params[:id])
+    else
+      redirect_to edit_user_path(@current_user.id)
+    end
   end
 
   def update
     user = User.find(params[:id])
-    
+
     if user.update user_params
        redirect_to user_path
     else
       flash[:errors] = user.errors
       redirect_to edit_user_path
-     
     end
   end
 
   def destroy
-    user = User.find(params[:id])
-    user.destroy
+    # To prevent rogue post delete request and only allow the actual user to delete their profile
+    if @current_user.id == params[:id].to_i
+      user = User.find(params[:id])
+      user.destroy
+      session[:destination_id] = nil
+    end
     redirect_to root_path
   end
